move gene utils into gene file
