+++ seqr/views/utils/orm_to_json_utils_tests.py
@@ -7,8 +7,8 @@ from seqr.models import Project, Family, Individual, Sample, IgvSample, SavedVar
     VariantNote, LocusList, VariantSearch
 from seqr.views.utils.orm_to_json_utils import _get_json_for_user, _get_json_for_project, _get_json_for_family, \
     _get_json_for_individual, get_json_for_sample, get_json_for_saved_variant, get_json_for_variant_tags, \
-    get_json_for_variant_functional_data_tags, get_json_for_variant_note, get_json_for_locus_list, get_json_for_gene, \
-    get_json_for_saved_search, get_json_for_saved_variants_with_tags, get_json_for_genes
+    get_json_for_variant_functional_data_tags, get_json_for_variant_note, get_json_for_locus_list, \
+ get_json_for_saved_search, get_json_for_saved_variants_with_tags
 from seqr.views.utils.test_utils import USER_FIELDS, PROJECT_FIELDS, FAMILY_FIELDS, INTERNAL_FAMILY_FIELDS, \
     INDIVIDUAL_FIELDS, INTERNAL_INDIVIDUAL_FIELDS, INDIVIDUAL_FIELDS_NO_FEATURES, SAMPLE_FIELDS, SAVED_VARIANT_FIELDS,  \
     FUNCTIONAL_FIELDS, SAVED_SEARCH_FIELDS, LOCUS_LIST_DETAIL_FIELDS, GENE_FIELDS, GENE_DETAIL_FIELDS, IGV_SAMPLE_FIELDS, \
@@ -186,33 +186,3 @@ class JSONUtilsTest(TestCase):
         user = User.objects.filter().first()
         json = get_json_for_locus_list(locus_list, user)
         self.assertSetEqual(set(json.keys()), LOCUS_LIST_DETAIL_FIELDS)
-
-    def test_json_for_gene(self):
-        gene = GeneInfo.objects.get(id=1)
-        json = get_json_for_gene(gene, user=None)
-        self.assertSetEqual(set(json.keys()), GENE_DETAIL_FIELDS)
-
-    def test_json_for_genes(self):
-        gene_ids = {'ENSG00000223972', 'ENSG00000227232'}
-        genes = GeneInfo.objects.filter(gene_id__in=gene_ids)
-        user = User.objects.filter().first()
-
-        json = get_json_for_genes(genes, user)
-        self.assertSetEqual({gene['geneId'] for gene in json}, gene_ids)
-        self.assertSetEqual(set(json[0].keys()), GENE_FIELDS)
-
-        fields = {'constraints', 'omimPhenotypes', 'mimNumber', 'cnSensitivity'}
-        fields.update(GENE_FIELDS)
-        json = get_json_for_genes(genes, user, add_variant_gene_display_fields=True)
-        self.assertSetEqual({gene['geneId'] for gene in json}, gene_ids)
-        self.assertSetEqual(set(json[0].keys()), fields)
-
-        json = get_json_for_genes(genes, user, add_variant_gene_fields=True)
-        self.assertSetEqual({gene['geneId'] for gene in json}, gene_ids)
-        self.assertSetEqual(set(json[0].keys()), GENE_VARIANT_FIELDS)
-
-        json = get_json_for_genes(genes, user, add_all=True)
-        self.assertSetEqual({gene['geneId'] for gene in json}, gene_ids)
-        self.assertSetEqual(set(json[0].keys()), GENE_DETAIL_FIELDS)
-
-
